+++ spec/dummy/app/controllers/inertia_test_controller.rb
@@ -53,6 +53,10 @@ class InertiaTestController < ApplicationController
     redirect_to 'http://external-website.com/some_path', allow_other_host: true
   end
 
+  def full_page_redirect_test
+    redirect_to empty_test_path, inertia: { full_page: true }
+  end
+
   def location_header_test
     head (params[:status] || 302).to_i, location: params[:url]
   end